Move Embellisher to own file
